third_party/blink/web_tests/external/wpt/dom/collections/HTMLCollection-as-prototype-expected.txt

This is a testharness.js-based test.
PASS HTMLCollection as a prototype should not allow getting .length on the base object
FAIL HTMLCollection as a prototype and setting own properties assert_equals: expected (string) "foo" but got (object) Element node <p id="named"></p>
Harness: the test ran to completion.
